Improved wasm support
This commit adds a feature `no-free` and a patch that makes QuickJS' `__JS_FreeValueRT` be a noop. 

This is useful in certain short-lived Wasm programs, where freeing values in the runtime itself is not a concern and it can be assumed that the Wasm runtime will clean up any memory by recycling the Instance's memory at the end of the execution.